refactor: rename main entry file to index.ts and remove unused ResizeManager and main.ts files
