+++ src/main.ts
@@ -1,2 +1,2 @@
 import './style.css'
-import '../lib/main'
+import '../lib/index'
