+++ vite.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vite'
 export default defineConfig({
   build: {
     lib: {
-      entry: './lib/main.ts',
+      entry: './lib/index.ts',
       name: 'twind-scope',
       fileName: (format) =>
         `twind-scope.${
